requirements.txt
openai-whisper
pytest
sounddevice
vosk
